IP2-545: return to one new event component
